fix: issues with disconnecting and completing
Part of the DDB refactor messed up which keys we needed to remove a record. Upon investigation we had a "topic" as a range key because we used to allow multiple topics per subscription. I'm pulling that out for the time being.

- add more logging and fixup the logging function's input
- Better DDB logging
- The logger function now always gets the object, the type change is additive so it's not a breaking change (I kind of want to move pino for logging and add a concept of log level)
- DDB can now support range keys even though we don't use them

BREAKING CHANGE: The subscriptions Table has has its range key removed. This will require a migration.