fix: remove ESM (#304)
vitest, vite, and a few other bundlers are having issues with the graphql subpath imports as they're not declared in the graphql export map. This is fine for commonjs but it wreaks havock on these esm loaders. I think this is a bug in the commonjs/esm interop but here we are.

BREAKING CHANGE: This removes the ESM export from the export maps and the shipped files to work around buggy esm/commonjs interop